refactor(interface): added the interface module.
