+++ src/stimulus/data/interface/__init__.py
@@ -1,0 +1,1 @@
+"""Module for interfacing configs with the pipeline."""
